Move event types and sub types to one file
