+++ src/app/(frontend)/[center]/events/page.tsx
@@ -3,8 +3,7 @@ import type { Metadata, ResolvedMetadata } from 'next/types'
 import { getPayload, Where } from 'payload'
 import { Suspense } from 'react'
 
-import { eventSubTypesData } from '@/collections/Events/components/eventSubTypes'
-import { eventTypesData } from '@/collections/Events/components/eventTypes'
+import { eventSubTypesData, eventTypesData } from '@/collections/Events/constants'
 import { EventCollection } from '@/components/EventCollection'
 import { PageRange } from '@/components/PageRange'
 import { Pagination } from '@/components/Pagination'
